backend/api/requirements.txt
# MAMA-LENS AI — Production Requirements (Python 3.11)

# ─── Web Framework ───────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0

# ─── Database — MongoDB ──────────────────────────────────────────────────────
motor==3.4.0
pymongo==4.7.2

# ─── Auth & Security ─────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
cryptography==42.0.7

# ─── Validation ──────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ─── AI / ML ─────────────────────────────────────────────────────────────────
numpy==1.26.4
scikit-learn==1.5.0

# ─── Utilities ───────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pytz==2024.1
structlog==24.1.0
